+++ RELEASES.md
@@ -8,6 +8,8 @@
 
 ### 0.2.0 Changed
 
+- renamed `taffy::forest::Forest.new-node(..)` `taffy::forest::Forest.new_with_children(..)`
+- renamed `taffy::node::Taffy.new-node(..)` -> `taffy::node::Taffy.new_with_children(..)`
 - renamed `taffy::style::Style` -> `taffy::style::FlexboxLayout` to more precicely indicate its purpose
 
 ### 0.2.0 Fixed